requirements.txt
Django==1.4.5
PyJWT==1.3.0
South==1.0
argparse==1.2.1
django-registration==1.0
django-tinymce==1.5.3
httplib2==0.9
ipaddr==2.1.11
longerusername==0.4
lxml==3.4.0
oauth2==1.5.211
oauthlib==0.6.3
python-openid==2.2.5
python-social-auth==0.2.10
requests==2.7.0
requests-oauthlib==0.4.2
six==1.8.0
wsgiref==0.1.2
